Filter: fix unit tests and many warnings

+++ src/filter/operations/float.c
@@ -5,7 +5,7 @@
  * \date 2020
  */
 
-/* 
+/*
  * Copyright (C) 2020 CESNET, z.s.p.o.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -42,6 +42,7 @@
 
 #include "float.h"
 #include "../values.h"
+#include <math.h>
 
 #define FLOAT_EQUALS_EPSILON   0.001 // Precision to 3 decimal point places when comparing floats for equality
 
@@ -139,7 +140,7 @@ void
 float_in_list(fds_filter_value_u *item, fds_filter_value_u *list, fds_filter_value_u *result)
 {
     result->b = false;
-    for (int i = 0; i < list->list.len; i++) {
+    for (uint64_t i = 0; i < list->list.len; i++) {
         if (list->list.items[i].f == item->f) {
             result->b = true;
             return;
